Adding New Class for better readability
